projeto_ETL_Santander/requirements.txt
pandas
numpy
faker
matplotlib
seaborn
pyyaml
typing